+++ templates/parts/block/pattern-features-steps.php
@@ -1,0 +1,112 @@
+<?php
+/**
+ * Block pattern setup file.
+ *
+ * Default pattern categories: buttons, columns, gallery, header, text.
+ *
+ * @package    Michelle
+ * @copyright  WebMan Design, Oliver Juhas
+ *
+ * @since  1.0.10
+ */
+
+namespace WebManDesign\Michelle\Content;
+
+// Exit if accessed directly.
+defined( 'ABSPATH' ) || exit;
+
+// Add block pattern setup args.
+Block_Patterns::add_pattern_args( __FILE__, array(
+	'title'      => _x( 'Features: Steps', 'Block pattern title.', 'michelle' ),
+	'categories' => array( 'columns' ),
+	'keywords'   => array( esc_html__( 'services', 'michelle' ) ),
+) );
+
+// Block pattern content:
+
+$image = Starter::get_image_url( '16to9' );
+
+?>
+
+<!-- wp:cover {"url":"<?php echo esc_url_raw( $image ); ?>","id":999999,"dimRatio":90,"overlayColor":"accent","align":"full","className":"is-style-no-padding"} -->
+<div class="wp-block-cover alignfull has-background-dim-90 has-accent-background-color has-background-dim is-style-no-padding"><img class="wp-block-cover__image-background wp-image-999999" alt="" src="<?php echo esc_url_raw( $image ); ?>" data-object-fit="cover"/><div class="wp-block-cover__inner-container"><!-- wp:spacer {"height":150} -->
+<div style="height:150px" aria-hidden="true" class="wp-block-spacer"></div>
+<!-- /wp:spacer -->
+
+<!-- wp:columns {"align":"wide"} -->
+<div class="wp-block-columns alignwide"><!-- wp:column -->
+<div class="wp-block-column"><!-- wp:html -->
+<!-- https://iconify.design/icon-sets/ic/outline-filter-1.html -->
+<svg class="svg-icon" width="5em" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" viewBox="0 0 24 24">
+<path d="M3 5H1v16c0 1.1.9 2 2 2h16v-2H3V5zm11 10h2V5h-4v2h2v8zm7-14H7c-1.1 0-2 .9-2 2v14c0 1.1.9 2 2 2h14c1.1 0 2-.9 2-2V3c0-1.1-.9-2-2-2zm0 16H7V3h14v14z" fill="#ffffff"></path>
+</svg>
+<!-- /wp:html -->
+
+<!-- wp:spacer {"height":50} -->
+<div style="height:50px" aria-hidden="true" class="wp-block-spacer"></div>
+<!-- /wp:spacer -->
+
+<!-- wp:paragraph {"className":"is-style-no-margin-vertical","fontSize":"large"} -->
+<p class="is-style-no-margin-vertical has-large-font-size">Lorem ipsum dolor sit amet  elit laoreet, nostra dui nulla.</p>
+<!-- /wp:paragraph -->
+
+<!-- wp:paragraph {"fontSize":"extra-large"} -->
+<p class="has-extra-large-font-size">→</p>
+<!-- /wp:paragraph --></div>
+<!-- /wp:column -->
+
+<!-- wp:column {"verticalAlignment":"center"} -->
+<div class="wp-block-column is-vertically-aligned-center"><!-- wp:spacer {"height":70} -->
+<div style="height:70px" aria-hidden="true" class="wp-block-spacer"></div>
+<!-- /wp:spacer -->
+
+<!-- wp:html -->
+<!-- https://iconify.design/icon-sets/ic/outline-filter-2.html -->
+<svg class="svg-icon" width="5em" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" viewBox="0 0 24 24">
+<path d="M3 5H1v16c0 1.1.9 2 2 2h16v-2H3V5zm18-4H7c-1.1 0-2 .9-2 2v14c0 1.1.9 2 2 2h14c1.1 0 2-.9 2-2V3c0-1.1-.9-2-2-2zm0 16H7V3h14v14zm-4-4h-4v-2h2a2 2 0 0 0 2-2V7a2 2 0 0 0-2-2h-4v2h4v2h-2a2 2 0 0 0-2 2v4h6v-2z" fill="#ffffff"></path>
+</svg>
+<!-- /wp:html -->
+
+<!-- wp:spacer {"height":50} -->
+<div style="height:50px" aria-hidden="true" class="wp-block-spacer"></div>
+<!-- /wp:spacer -->
+
+<!-- wp:paragraph {"className":"is-style-no-margin-vertical","fontSize":"large"} -->
+<p class="is-style-no-margin-vertical has-large-font-size">Lorem ipsum dolor sit amet  elit laoreet, nostra dui nulla.</p>
+<!-- /wp:paragraph -->
+
+<!-- wp:paragraph {"fontSize":"extra-large"} -->
+<p class="has-extra-large-font-size">→</p>
+<!-- /wp:paragraph -->
+
+<!-- wp:spacer {"height":70} -->
+<div style="height:70px" aria-hidden="true" class="wp-block-spacer"></div>
+<!-- /wp:spacer --></div>
+<!-- /wp:column -->
+
+<!-- wp:column {"verticalAlignment":"bottom"} -->
+<div class="wp-block-column is-vertically-aligned-bottom"><!-- wp:html -->
+<!-- https://iconify.design/icon-sets/ic/outline-filter-3.html -->
+<svg class="svg-icon" width="5em" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" viewBox="0 0 24 24">
+<path d="M21 1H7c-1.1 0-2 .9-2 2v14c0 1.1.9 2 2 2h14c1.1 0 2-.9 2-2V3c0-1.1-.9-2-2-2zm0 16H7V3h14v14zM3 5H1v16c0 1.1.9 2 2 2h16v-2H3V5zm14 8v-1.5c0-.83-.67-1.5-1.5-1.5c.83 0 1.5-.67 1.5-1.5V7a2 2 0 0 0-2-2h-4v2h4v2h-2v2h2v2h-4v2h4a2 2 0 0 0 2-2z" fill="#ffffff"></path>
+</svg>
+<!-- /wp:html -->
+
+<!-- wp:spacer {"height":50} -->
+<div style="height:50px" aria-hidden="true" class="wp-block-spacer"></div>
+<!-- /wp:spacer -->
+
+<!-- wp:paragraph {"className":"is-style-no-margin-vertical","fontSize":"large"} -->
+<p class="is-style-no-margin-vertical has-large-font-size">Lorem ipsum dolor sit amet  elit laoreet, nostra dui nulla.</p>
+<!-- /wp:paragraph -->
+
+<!-- wp:paragraph {"fontSize":"extra-large"} -->
+<p class="has-extra-large-font-size">→</p>
+<!-- /wp:paragraph --></div>
+<!-- /wp:column --></div>
+<!-- /wp:columns -->
+
+<!-- wp:spacer {"height":150} -->
+<div style="height:150px" aria-hidden="true" class="wp-block-spacer"></div>
+<!-- /wp:spacer --></div></div>
+<!-- /wp:cover -->